requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
psycopg==3.2.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
python-dotenv~=1.0.1
graphene-django~=3.2.2
graphene~=3.4.3